DINC0468330:All the attachment is getting deleted when in create mode…

+++ sdm/src/main/java/com/sap/cds/sdm/handler/TokenHandler.java
@@ -310,7 +310,7 @@ public class TokenHandler {
     DefaultHttpClientFactory.DefaultHttpClientFactoryBuilder builder =
         DefaultHttpClientFactory.builder();
     if (connectionPoolConfig == null) {
-      Duration timeout = Duration.ofSeconds((long) SDMConstants.CONNECTION_TIMEOUT);
+      Duration timeout = Duration.ofSeconds(SDMConstants.CONNECTION_TIMEOUT);
       builder.timeoutMilliseconds((int) timeout.toMillis());
       builder.maxConnectionsPerRoute(SDMConstants.MAX_CONNECTIONS);
       builder.maxConnectionsTotal(SDMConstants.MAX_CONNECTIONS);
